Update README.md
The aim of this analysis is to model and forecast the trend of guest occupancy rate of a hote; in Abuja, Nigeria. To examine factors that facilitate the increment and sustainability of occupancy levels in the hotel. The year under review is 2013-2021. Autoregressive Integrated Moving Average (ARIMA) Model was used to model the occupancy rate with the 4-year forecast.

The data was collected from the organisation. R Programming Language with readr, dplyr , tseries, forecast libraries.

Steps taken for the ARIMA Model *Visualize the time series! *Stationarize the time series *Plot ACF/PACF and find optimal parameters! *Build the ARIMA model! *Make predictions!

This gave us the best ARIMA Model to be Best model: ARIMA(5,0,0) with minimum AKAIKE INFORMATION CRITERIA (AIC) value of 102.818. Checking model 5,0,0 gave us the intercept of 0.02 and AIC of 80.21. PACF residual shows a better chart with normal lags. The histogram shows, it is normally distributed

Finally, I concluded that The majority of researchers have conducted extensive research on forecasting room occupancy and attendance in hotels in various countries using various approaches. The rates are by definition noisy, non-stationary, and deterministically chaotic. Many economic, political, and psychological factors influence room occupancy and attendance rate in Hotels, some of these factors identified in this study include quality of effective service delivery, finance, competition, tourism and infrastructure. Forecasting room occupancy rates is not a simple task because of these unstable factors. Wrong adjustments which has an adverse effect on the decision making process.

Thank you for your review